# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: Bankcard API made as a test task
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/cards/create:
    post:
      tags:
        - Admin - Cards
      summary: Create card
      description: Creates a card with given currency for given user
      operationId: createCard
      requestBody:
        description: Card creation parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreationDto"
        required: true
      responses:
        "200":
          description: Card was created and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
...
  /api/admin/cards/status:
    patch:
      tags:
        - Admin - Cards
      summary: Update card status
      description: Changes status of the card with given id to a desired one
      operationId: updateCardStatus
      requestBody:
        description: Card status update parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardStatusUpdateDto"
        required: true
      responses:
        "200":
          description: Card's status was updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "404":
          description: Card with given id not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cards/search:
    get:
      tags:
        - User - Cards
      summary: Search cards with optional filters
      description: "Returns a paginated list of cards based on optional filters: balance,\
        \ comparison type, owner ID, and status"
      operationId: searchCards
      parameters:
        - name: balance
          in: query
          description: Balance to compare
          required: false
          schema:
            type: number
            format: double
          example: 1000.0
        - name: isGreater
          in: query
          description: Whether balance should be greater than the given value (true/false)
          required: false
          schema:
            type: boolean
          example: true
        - name: status
          in: query
          description: Card status to filter by
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - LOCKED
              - EXPIRED
              - PENDING_LOCK
      responses:
        "200":
          description: Card's pagination was returned successfully filled with CardDtos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/cards/search:
    get:
      tags:
        - Admin - Cards
      summary: Search cards with optional filters
      description: "Returns a paginated list of cards based on optional filters: balance,\
        \ comparison type, owner ID, and status"
      operationId: searchCards_1
      parameters:
        - name: balance
          in: query
          description: Balance to compare
          required: false
          schema:
            type: number
            format: double
          example: 1000.0
        - name: isGreater
          in: query
          description: Whether balance should be greater than the given value (true/false)
          required: false
          schema:
            type: boolean
          example: true
        - name: ownerId
          in: query
          description: UUID of the card owner
          required: false
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        - name: status
          in: query
          description: Card status to filter by
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - LOCKED
              - EXPIRED
              - PENDING_LOCK
      responses:
        "200":
          description: Card's pagination was returned successfully filled with CardDtos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/cards/get/{id}:
    get:
      tags:
        - Admin - Cards
      summary: Get card by id
      description: Returns card dto with given id
      operationId: getCardById
      parameters:
        - name: id
          in: path
          description: UUID of the card
          required: true
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        "200":
          description: Card was found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/cards/delete/{id}:
    delete:
      tags:
        - Admin - Cards
      summary: Delete card by id
      description: Deletes card with given id
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          description: UUID of the card
          required: true
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        "204":
          description: Card was deleted
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    CardDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - ACTIVE
            - LOCKED
            - EXPIRED
            - PENDING_LOCK
        currency:
          type: string
          enum:
            - RUB
            - USD
            - EUR
        balance:
          type: number
        masked_number:
          type: string
        owner_id:
          type: string
          format: uuid
        expiration_date:
          type: string
          format: date
    ContentDisposition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        charset:
          type: string
        size:
          type: integer
          format: int64
          deprecated: true
        creationDate:
          type: string
          format: date-time
          deprecated: true
        modificationDate:
          type: string
          format: date-time
          deprecated: true
        readDate:
          type: string
          format: date-time
          deprecated: true
        attachment:
          type: boolean
        formData:
          type: boolean
        inline:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        body:
          $ref: "#/components/schemas/ProblemDetail"
        headers:
          type: object
          properties:
            connection:
              type: array
              items:
                type: string
            contentDisposition:
              $ref: "#/components/schemas/ContentDisposition"
            acceptCharset:
              type: array
              items:
                type: string
            location:
              type: string
              format: uri
            empty:
              type: boolean
            host:
              type: object
              properties:
                hostString:
                  type: string
                address:
                  type: object
                  properties:
                    hostAddress:
                      type: string
                    address:
                      type: array
                      items:
                        type: string
                        format: byte
                    hostName:
                      type: string
                    linkLocalAddress:
                      type: boolean
                    multicastAddress:
                      type: boolean
                    anyLocalAddress:
                      type: boolean
                    loopbackAddress:
                      type: boolean
                    siteLocalAddress:
                      type: boolean
                    mcglobal:
                      type: boolean
                    mcnodeLocal:
                      type: boolean
                    mclinkLocal:
                      type: boolean
                    mcsiteLocal:
                      type: boolean
                    mcorgLocal:
                      type: boolean
                    canonicalHostName:
                      type: string
                port:
                  type: integer
                  format: int32
                unresolved:
                  type: boolean
                hostName:
                  type: string
            all:
              type: object
              additionalProperties:
                type: string
              writeOnly: true
            lastModified:
              type: integer
              format: int64
            date:
              type: integer
              format: int64
            contentLength:
              type: integer
              format: int64
            contentType:
              $ref: "#/components/schemas/MediaType"
            ifModifiedSince:
              type: integer
              format: int64
            origin:
              type: string
            range:
              type: array
              items:
                $ref: "#/components/schemas/HttpRange"
            allow:
              uniqueItems: true
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            cacheControl:
              type: string
            contentLanguage:
              type: object
              properties:
                language:
                  type: string
                displayName:
                  type: string
                country:
                  type: string
                variant:
                  type: string
                script:
                  type: string
                unicodeLocaleAttributes:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                unicodeLocaleKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                displayLanguage:
                  type: string
                displayScript:
                  type: string
                displayCountry:
                  type: string
                displayVariant:
                  type: string
                extensionKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                iso3Language:
                  type: string
                iso3Country:
                  type: string
            etag:
              type: string
            accessControlAllowOrigin:
              type: string
            accessControlAllowMethods:
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            accessControlAllowHeaders:
              type: array
              items:
                type: string
            accessControlExposeHeaders:
              type: array
              items:
                type: string
            accessControlAllowCredentials:
              type: boolean
            accessControlMaxAge:
              type: integer
              format: int64
            accessControlRequestMethod:
              $ref: "#/components/schemas/HttpMethod"
            accessControlRequestHeaders:
              type: array
              items:
                type: string
            accept:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            acceptLanguageAsLocales:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                  displayName:
                    type: string
                  country:
                    type: string
                  variant:
                    type: string
                  script:
                    type: string
                  unicodeLocaleAttributes:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  unicodeLocaleKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  displayLanguage:
                    type: string
                  displayScript:
                    type: string
                  displayCountry:
                    type: string
                  displayVariant:
                    type: string
                  extensionKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  iso3Language:
                    type: string
                  iso3Country:
                    type: string
            acceptPatch:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            bearerAuth:
              type: string
              writeOnly: true
            expires:
              type: integer
              format: int64
            ifMatch:
              type: array
              items:
                type: string
            ifNoneMatch:
              type: array
              items:
                type: string
            ifUnmodifiedSince:
              type: integer
              format: int64
            pragma:
              type: string
            upgrade:
              type: string
            vary:
              type: array
              items:
                type: string
            acceptLanguage:
              type: array
              items:
                type: object
                properties:
                  range:
                    type: string
                  weight:
                    type: number
                    format: double
            basicAuth:
              type: string
              writeOnly: true
          additionalProperties:
            type: array
            items:
              type: string
        statusCode:
          $ref: "#/components/schemas/HttpStatusCode"
        typeMessageCode:
          type: string
        detailMessageCode:
          type: string
        detailMessageArguments:
          type: array
          items:
            type: object
        titleMessageCode:
          type: string
    HttpMethod:
      type: object
    HttpRange:
      type: object
    HttpStatusCode:
      type: object
      properties:
        is4xxClientError:
          type: boolean
        error:
          type: boolean
        is5xxServerError:
          type: boolean
        is1xxInformational:
          type: boolean
        is2xxSuccessful:
          type: boolean
        is3xxRedirection:
          type: boolean
    MediaType:
      type: object
      properties:
        type:
          type: string
        subtype:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        qualityValue:
          type: number
          format: double
        wildcardType:
          type: boolean
        wildcardSubtype:
          type: boolean
        subtypeSuffix:
          type: string
        charset:
          type: string
        concrete:
          type: boolean
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    CardCreationDto:
      required:
        - currency
        - user_id
      type: object
      properties:
        currency:
          type: string
          enum:
            - RUB
            - USD
            - EUR
        user_id:
          type: string
          format: uuid
...
    CardStatusUpdateDto:
      required:
        - card_id
        - status
      type: object
      properties:
        status:
          type: string
          enum:
            - ACTIVE
            - LOCKED
            - EXPIRED
            - PENDING_LOCK
        card_id:
          type: string
          format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT